+++ NEWS.md
@@ -1,4 +1,4 @@
-# pathfindR 1.2.9001
+# pathfindR 1.2.9002
 To be released as 1.2.1
 
 ## Major Changes
@@ -10,6 +10,9 @@ To be released as 1.2.1
 - in `calculate_pw_scores`, if `cases` are provided, the pathways are reordered before plotting the heat map and returning the matrix according to their activity in `cases`. This way, "up" pathways are grouped together, same for "down" pathways.
 - in `calculate_pwd`, if a pathway has perfect overlap with other pathways, change the correlation value with 1 instead of NA.
 - in `choose_clusters`, if `result_df` has less than 3 pathways, do not perform clustering.
+- `run_pathfindR` checks whether the output directory (`output_dir`) already exists and if it exists, now appends "(1)" to `output_dir` and displays a warning message. This was implemented to prevent writing over existing results.
+- in run `run_pathfindR`, recursive creation for the output directory (`output_dir`) is now supported.
+- in run `run_pathfindR`, if no pathways are found, the function returns an empty data frame instead of raising an error.
 
 # pathfindR 1.2